test(property): extract property tests to dedicated file + add scenario docstrings
- Move TestDescriptorOrderAgnostic, TestPropertyOrderAgnostic, TestPropertyErrorPaths (624 lines) from test_deprecation.py to tests/unittests/test_property.py
- Remove no-longer-needed imports from test_deprecation.py: types, Optional, cached_property, validate_deprecation_expiry, InnerOrderDeprecatedPropCls, DelOnlyDeprecatedPropCls, _DeprecatedProperty
- Add prose-paragraph scenario descriptions to TestPropertyErrorPaths test docstrings
- Require scenario description prose paragraph in all non-trivial test method docstrings (AGENTS.md + CONTRIBUTING.md)

---
